+++ src/entity/config.ts
@@ -1,4 +1,4 @@
-import { ConfigurationChangeEvent, workspace } from "vscode"
+import { ConfigurationChangeEvent, workspace, window, ColorThemeKind } from "vscode"
 
 const SECTION = "crates-cmp"
 const SPARSE_INDEX_CONFIG = "crates.sparse-index.url"
@@ -7,10 +7,12 @@ const ENABLE_CARGO_AUDIT = "cargo.audit.enable"
 class Config {
     private sparseIndexUrl: string = "https://index.crates.io"
     private enableAudit: boolean = false
+    private currentTheme: ColorThemeKind
 
     constructor() {
         this.updateSparseIndexUrl()
         this.updateCargoAudit()
+        this.currentTheme = window.activeColorTheme.kind
     }
 
     getSparseIndexUrl() {
@@ -21,6 +23,10 @@ class Config {
         return this.enableAudit
     }
 
+    getCurrentTheme() {
+        return this.currentTheme
+    }
+
     updateSparseIndexUrl() {
         const sparseIndexUrl = workspace.getConfiguration("crates-cmp").get(SPARSE_INDEX_CONFIG)
         if (typeof sparseIndexUrl === "string") {
@@ -29,11 +35,16 @@ class Config {
     }
 
     updateCargoAudit() {
-        const enableAudit = workspace.getConfiguration("crates-cmp").get(ENABLE_CARGO_AUDIT)
+        const enableAudit = workspace.getConfiguration("workbench.colorTheme").get(ENABLE_CARGO_AUDIT)
         if (typeof enableAudit === "boolean") {
             this.enableAudit = enableAudit
         }
     }
+
+    onThemeChange(theme) {
+        this.currentTheme = theme.kind
+    }
+
 
     onChange(e: ConfigurationChangeEvent) {
         if (e.affectsConfiguration(SECTION + "." + SPARSE_INDEX_CONFIG)) {